experiments/text-summarizer/requirements.txt
google-genai==0.3.0